WIP on saltcloud
/cc @toutpt @regilero @jpcw @guitdz

near the goal, now i can provision lxcs and they are attached to my master
with the salt-cloud cli tool.

I now need to stress test and integrate them with the cloud.present states
which are coded but not tested...

Please note that this include also the saltify driver to handle baremetal servers.
So i hope that without any further surprise before late wednesday this will be a finished topic.