makina-states lxc releaser
/cc @regilero: with this function, we can upload to sourceforce the tarball to serve
as a base for the salt cloud controller.
As the image is a bit long and picky to construct, it is a bad idea to redo it
from scratch at anytime.
So i am adopting the same scheme than with vms to keep an uptodate tarball
somewhere.

From there the mastersalt cloud controller sync its local copy (the extraction
result) on the other nodes via rsync. (see the salt-runner i added)