Full stack SSL from client to backends
This make haproxy breaks ssl but also speak ssl with backends, only if
available and fallback on http if the backend does not respond over ssl.

This is experimental for now, and will be deployed progressivly on our
cluster. But by chance, we will be more and more saykioure ;).

cc @chabby @sebcorbin @jpcw @regilero